+++ src/redis-cpu-panel/components/index.ts
@@ -1,0 +1,2 @@
+export * from './redis-cpu-panel-graph';
+export * from './redis-cpu-panel';