+++ src/redis-cpu-panel/types.ts
@@ -1,0 +1,88 @@
+import { DataQuery, DateTime } from '@grafana/data';
+
+/**
+ * Panel Options
+ */
+export interface PanelOptions {
+  /**
+   * Interval
+   *
+   * @type {number}
+   */
+  interval: number;
+
+  /**
+   * Max Items
+   *
+   * @type {number}
+   */
+  maxItemsPerSeries: number;
+}
+
+/**
+ * Redis Query
+ */
+export interface RedisQuery extends DataQuery {
+  /**
+   * Query command
+   *
+   * @type {string}
+   */
+  query?: string;
+
+  /**
+   * Redis Command type
+   *
+   * @type {string}
+   */
+  type?: string;
+
+  /**
+   * Redis Command
+   *
+   * @type {string}
+   */
+  command?: string;
+
+  /**
+   * Redis Section
+   *
+   * @type {string}
+   */
+  section?: string;
+}
+
+/**
+ * Object which keeps SeriesValue[] by command name
+ */
+export interface SeriesMap {
+  [key: string]: SeriesValue[];
+}
+
+/**
+ * Series Value
+ */
+export interface SeriesValue {
+  time: DateTime;
+  value: number;
+}
+
+/**
+ * Calculation
+ */
+export interface ValuesForCalculation {
+  /**
+   * Time
+   */
+  time: DateTime;
+
+  /**
+   * User
+   */
+  user: number;
+
+  /**
+   * System
+   */
+  system: number;
+}